Add functions to MapHandler
Complete the working of `enter_net`: after retrieve data from the
outside send data from inside to the outside.

Add function `set_participant` to handle propagation of
participation to a optional service.

TODO: we need to the delegate `GetBroadcastNeighbors` the parameter
of type `IPeersMissingArcHandler` in order to specify what to do
when an arc does not confirm the receipt.

TODO: we need to move utility functions (e.g. `make_tuple_gnode`,
`convert_tuple_gnode`) to another source file, outside from
the class.